api/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
sqlalchemy==2.0.27
PyMySQL==1.1.0
cryptography==42.0.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
redis==5.0.1
python-dotenv==1.0.1
pydantic==2.6.1
pydantic[email]==2.6.1
alembic==1.13.1
python-multipart>=0.0.6
jinja2>=3.1.2
email-validator>=2.0.0